+++ modules/perception/common/graph/gated_hungarian_bigraph_matcher.h
@@ -192,8 +192,8 @@ void GatedHungarianMatcher<T>::ComputeConnectedComponents(
 
   std::vector<std::vector<int>> nb_graph;
   nb_graph.resize(rows_num_ + cols_num_);
-  for (int i = 0; i < rows_num_; ++i) {
-    for (int j = 0; j < cols_num_; ++j) {
+  for (size_t i = 0; i < rows_num_; ++i) {
+    for (size_t j = 0; j < cols_num_; ++j) {
       if (is_valid_cost_(global_costs_(i, j))) {
         nb_graph[i].push_back(static_cast<int>(rows_num_) + j);
         nb_graph[j + rows_num_].push_back(i);
@@ -210,7 +210,7 @@ void GatedHungarianMatcher<T>::ComputeConnectedComponents(
   for (size_t i = 0; i < components.size(); ++i) {
     for (size_t j = 0; j < components[i].size(); ++j) {
       int id = components[i][j];
-      if (id < rows_num_) {
+      if (id < static_cast<int>(rows_num_)) {
         row_components->at(i).push_back(id);
       } else {
         id -= static_cast<int>(rows_num_);
